refactor: 공통 로직 최소화 및 컴포넌트 내부로 추가 및 삭제 코드 넣기
