+++ styles/thumbnail.css
@@ -7,6 +7,12 @@
   gap: 70px;
 }
 
+.skeleton-list {
+  margin: 0 auto 56px;
+  display: grid;
+  grid-template-columns: repeat(5, 200px);
+  gap: 70px;
+}
 .thumbnail {
   width: 200px;
   height: 300px;
